agregar estados de usuario

+++ src/main/java/com/perfulandia/usuarios/model/entity/TokenRecuperacion.java
@@ -1,15 +1,9 @@
 package com.perfulandia.usuarios.model.entity;
 
-import java.time.LocalDateTime;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
+import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "tokens_recuperacion")
@@ -18,7 +12,7 @@ import lombok.Setter;
 public class TokenRecuperacion {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, unique = true)
     private String token;
