+++ src/main/java/com/perfulandia/usuarios/repository/TokenRecuperacionRepository.java
@@ -1,11 +1,9 @@
 package com.perfulandia.usuarios.repository;
 
+import com.perfulandia.usuarios.model.entity.TokenRecuperacion;
+import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.perfulandia.usuarios.model.entity.TokenRecuperacion;
-
-public interface TokenRecuperacionRepository extends JpaRepository<TokenRecuperacion, Integer> {
+public interface TokenRecuperacionRepository extends JpaRepository<TokenRecuperacion, Long> {
     Optional<TokenRecuperacion> findByToken(String token);
 }
